# internal/api/handlers_points.go
package api

import (
	"fmt"
	"net/http"
	"strconv"
	"strings"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/nodelistdb/internal/database"
	"github.com/nodelistdb/internal/storage"
)

// resolvePointDomain picks the FTN network for a point endpoint. It consults
// the points table itself rather than the boss node: resolving via the node
// could 404 a point that exists only in a network the node-level heuristic
// does not prefer. A nil point resolves at boss level.
func (s *Server) resolvePointDomain(r *http.Request, zone, net, node int, point *int) (string, []string) {
	domains, err := s.storage.GetPointDomains(r.Context(), zone, net, node, point)
	if err != nil || domains == nil {
		// Non-nil so JSON bodies carry "available_domains": [] rather than null
		domains = []string{}
	}
	return preferDomain(explicitDomain(r), domains), domains
}

// parse4DPathParams extracts zone/net/node (and optionally point) from Chi
// URL parameters, writing the error response itself on failure.
func parse4DPathParams(w http.ResponseWriter, r *http.Request, withPoint bool) (zone, net, node, point int, ok bool) {
	for _, p := range []struct {
		name string
		dest *int
	}{
		{"zone", &zone}, {"net", &net}, {"node", &node},
	} {
		v, err := strconv.Atoi(chi.URLParam(r, p.name))
		if err != nil {
			WriteJSONError(w, fmt.Sprintf("Invalid %s parameter", p.name), http.StatusBadRequest)
			return 0, 0, 0, 0, false
		}
		*p.dest = v
	}
	if withPoint {
		v, err := strconv.Atoi(chi.URLParam(r, "point"))
		if err != nil {
			WriteJSONError(w, "Invalid point parameter", http.StatusBadRequest)
			return 0, 0, 0, 0, false
		}
		point = v
	}
	return zone, net, node, point, true
}

// SearchPointsHandler handles point search requests.
// GET /api/points?zone=2&net=5001&sysop_name=...&latest_only=true&limit=100
func (s *Server) SearchPointsHandler(w http.ResponseWriter, r *http.Request) {
	filter, hasConstraint, err := parsePointFilter(r)
	if err != nil {
		WriteJSONError(w, err.Error(), http.StatusBadRequest)
		return
	}

	if !hasConstraint {
		WriteJSONError(w, "Search requires at least one specific constraint (zone, net, node, point, system_name, location, sysop_name, list_source, or date range)", http.StatusBadRequest)
		return
	}

	points, err := s.storage.SearchPoints(r.Context(), filter)
	if err != nil {
		writeStorageErrorf(w, "Search failed", err)
		return
	}
	if points == nil {
		points = []database.Point{}
	}

	WriteJSONSuccess(w, pointSearchResponse{
		Points: points,
		Count:  len(points),
		Filter: pointFilterEcho{
			Zone: filter.Zone, Net: filter.Net, Node: filter.Node, Point: filter.PointNum,
			Domain: filter.Domain, ListSource: filter.ListSource,
			SystemName: filter.SystemName, Location: filter.Location, SysopName: filter.SysopName,
			DateFrom: filter.DateFrom, DateTo: filter.DateTo, LatestOnly: filter.LatestOnly,
			Limit: filter.Limit, Offset: filter.Offset,
		},
	})
}

// pointSearchResponse is the GET /api/points body.
type pointSearchResponse struct {
	Points []database.Point `json:"points"`
	Count  int              `json:"count"`
	Filter pointFilterEcho  `json:"filter"`
}

// pointFilterEcho is the search's reading of its parameters.
type pointFilterEcho struct {
	Zone       *int       `json:"zone"`
	Net        *int       `json:"net"`
	Node       *int       `json:"node"`
	Point      *int       `json:"point"`
	Domain     *string    `json:"domain"`
	ListSource *string    `json:"list_source"`
	SystemName *string    `json:"system_name"`
	Location   *string    `json:"location"`
	SysopName  *string    `json:"sysop_name"`
	DateFrom   *time.Time `json:"date_from"`
	DateTo     *time.Time `json:"date_to"`
	LatestOnly *bool      `json:"latest_only"`
	Limit      int        `json:"limit"`
	Offset     int        `json:"offset"`
}

// GetNodePointsHandler returns the snapshot points under a boss node.
// GET /api/nodes/{zone}/{net}/{node}/points?date=2024-01-01&domain=fidonet
func (s *Server) GetNodePointsHandler(w http.ResponseWriter, r *http.Request) {
	zone, net, node, _, ok := parse4DPathParams(w, r, false)
	if !ok {
		return
	}

	var asOf *time.Time
	if dateStr := r.URL.Query().Get("date"); dateStr != "" {
		d, err := time.Parse("2006-01-02", dateStr)
		if err != nil {
			WriteJSONError(w, "Invalid date format. Use YYYY-MM-DD", http.StatusBadRequest)
			return
		}
		asOf = &d
	}

	domain, availableDomains := s.resolvePointDomain(r, zone, net, node, nil)
	points, err := s.storage.GetPointsByBoss(r.Context(), domain, zone, net, node, asOf)
	if err != nil {
		writeStorageErrorf(w, "Failed to get points", err)
		return
	}
	if points == nil {
		points = []database.Point{}
	}

	response := addressEnvelope(zone, net, node, -1, domain, availableDomains)
	response["points"] = points
	response["count"] = len(points)
	if asOf != nil {
		response["as_of"] = asOf.Format("2006-01-02")
	}

	WriteJSONSuccess(w, response)
}

// GetPointHandler returns the current snapshot entry of one 4-D address.
// GET /api/points/{zone}/{net}/{node}/{point}
func (s *Server) GetPointHandler(w http.ResponseWriter, r *http.Request) {
	zone, net, node, point, ok := parse4DPathParams(w, r, true)
	if !ok {
		return
	}

	domain, availableDomains := s.resolvePointDomain(r, zone, net, node, &point)
	if len(availableDomains) > 1 {
		w.Header().Set("X-Available-Domains", strings.Join(availableDomains, ","))
	}

	latestOnly := true
	filter := database.PointFilter{
		Zone:       &zone,
		Net:        &net,
		Node:       &node,
		PointNum:   &point,
		Domain:     &domain,
		LatestOnly: &latestOnly,
		Limit:      1,
	}

	points, err := s.storage.SearchPoints(r.Context(), filter)
	if err != nil {
		writeStorageErrorf(w, "Point lookup failed", err)
		return
	}

	if len(points) == 0 {
		WriteJSONError(w, "Point not found", http.StatusNotFound)
		return
	}

	WriteJSONSuccess(w, points[0])
}

// GetPointHistoryHandler returns every stored entry of one 4-D address across
// all pointlist sources and dates.
// GET /api/points/{zone}/{net}/{node}/{point}/history
func (s *Server) GetPointHistoryHandler(w http.ResponseWriter, r *http.Request) {
	zone, net, node, point, ok := parse4DPathParams(w, r, true)
	if !ok {
		return
	}

	domain, availableDomains := s.resolvePointDomain(r, zone, net, node, &point)
	history, err := s.storage.GetPointHistory(r.Context(), domain, zone, net, node, point)
	if err != nil {
		writeStorageErrorf(w, "Failed to get point history", err)
		return
	}

	if len(history) == 0 {
		WriteJSONError(w, "Point not found", http.StatusNotFound)
		return
	}

	response := addressEnvelope(zone, net, node, point, domain, availableDomains)
	response["history"] = history
	response["count"] = len(history)

	WriteJSONSuccess(w, response)
}

// PointlistDatesHandler lists imported pointlist files (issue dates per
// series), newest first.
// GET /api/pointlists/dates?domain=fidonet&source=z2
func (s *Server) PointlistDatesHandler(w http.ResponseWriter, r *http.Request) {
	source := strings.ToLower(strings.TrimSpace(r.URL.Query().Get("source")))
	files, err := s.storage.GetPointlistDates(r.Context(), domainOrDefault(r), source)
	if err != nil {
		writeStorageErrorf(w, "Failed to get pointlist dates", err)
		return
	}

	if files == nil {
		files = []database.PointlistFile{}
	}
	response := map[string]interface{}{
		"files": files,
		"count": len(files),
	}

	WriteJSONSuccess(w, response)
}

// PointlistSourcesHandler summarizes the imported pointlist series.
// GET /api/pointlists/sources?domain=fidonet
func (s *Server) PointlistSourcesHandler(w http.ResponseWriter, r *http.Request) {
	sources, err := s.storage.GetPointlistSources(r.Context(), domainOrDefault(r))
	if err != nil {
		writeStorageErrorf(w, "Failed to get pointlist sources", err)
		return
	}

	if sources == nil {
		sources = []storage.PointlistSourceInfo{}
	}
	response := map[string]interface{}{
		"sources": sources,
		"count":   len(sources),
	}

	WriteJSONSuccess(w, response)
}