# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: NodelistDB API
  description: |
    NodelistDB is a comprehensive API for accessing historical FTN nodelist
    data: every FidoNet nodelist since 1986, other FTN networks (fsxNet, ...),
    pointlists (FTS-5002), and what the testing daemon found when it called
    each node.

    ## Features
    - **Node search**: find nodes by zone, net, node number, system name,
      location, sysop, capability flags or date range, in any network
    - **Historical data**: the complete history, change log and timeline of
      any node
    - **Multi-network**: `GET /api/networks` lists the imported FTN networks;
      most endpoints take `?domain=`
    - **Points**: pointlist snapshots under a boss, point history, imported
      pointlist issues
    - **Statistics**: per-date network statistics and the list of nodelist dates
    - **Reachability**: the testing daemon's results per node (BinkP, IFCICO,
      Telnet, FTP, VModem, over IPv4 and IPv6), trends and searches
    - **Software, hosting, PING/TRACE analytics**: distributions computed from
      those tests, and the netmail PING measurements
    - **Flag documentation**: a reference for FidoNet node flags

    ## Multi-network conventions
    A node address is only unique within one FTN network. Endpoints about one
    address resolve the network in this order: an explicit `?domain=`; the
    only network the address exists in; `fidonet` when it exists in several.
...
    as their descriptions say.

...

...

...
servers:
...
paths:
...
  /api/points:
    get:
      summary: Search Points
      description: |
        Search pointlist entries. Requires at least one constraint (zone, net,
        node, point, system_name, location, sysop_name, list_source, or date
        range). With latest_only=true the search applies snapshot semantics
        (one row per 4-D address, cross-source overlap resolved); otherwise
        every matching historical row is returned, newest first.
      operationId: searchPoints
      tags:
        - Points
      parameters:
        - name: zone
          in: query
          schema:
            type: integer
        - name: net
          in: query
          schema:
            type: integer
        - name: node
          in: query
          description: Boss node number
          schema:
            type: integer
        - name: point
          in: query
          description: Point number
          schema:
            type: integer
        - name: domain
          in: query
          description: FTN network (omit for all networks)
          schema:
            type: string
        - name: list_source
          in: query
          description: Pointlist series (r24, z2, ...)
          schema:
            type: string
        - name: system_name
          in: query
          description: System name (substring, case-insensitive, at least 2 characters)
          schema:
            type: string
        - name: location
          in: query
          description: Location (substring, case-insensitive, at least 2 characters)
          schema:
            type: string
        - name: sysop_name
          in: query
          description: Sysop name (substring, case-insensitive, at least 2 characters)
          schema:
            type: string
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          description: End of date range; with latest_only=true also the snapshot as-of date
          schema:
            type: string
            format: date
        - name: latest_only
          in: query
          description: Apply snapshot semantics (one row per 4-D address)
          schema:
            type: boolean
        - name: limit
          in: query
          description: Maximum number of results (default 100; values above 500 are capped to 500)
          schema:
            type: integer
            default: 100
            maximum: 500
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  points:
                    type: array
                    items:
                      $ref: '#/components/schemas/Point'
                  count:
                    type: integer
                  filter:
                    $ref: '#/components/schemas/PointSearchFilter'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/BudgetExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/points/{zone}/{net}/{node}/{point}/history:
    get:
      summary: Get Point History
      description: |
        Every stored entry of one 4-D point address across all pointlist
        series and issue dates. Rows are labeled with list_source;
        overlapping series are not collapsed.
      operationId: getPointHistory
      tags:
        - Points
      parameters:
        - $ref: '#/components/parameters/Zone'
        - $ref: '#/components/parameters/Net'
        - $ref: '#/components/parameters/Node'
        - $ref: '#/components/parameters/PointNumber'
        - $ref: '#/components/parameters/DomainResolved'
      responses:
        '200':
          description: Point history
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/AddressEnvelope'
                  - type: object
                    properties:
                      history:
                        type: array
                        items:
                          $ref: '#/components/schemas/Point'
                      count:
                        type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/BudgetExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/pointlists/dates:
    get:
      summary: List Imported Pointlist Files
      description: Imported pointlist issues (per series), newest first.
      operationId: getPointlistDates
      tags:
        - Points
      parameters:
        - $ref: '#/components/parameters/DomainDefaultFidonet'
        - name: source
          in: query
          description: Restrict to one pointlist series (r24, z2, ...)
          schema:
            type: string
      responses:
        '200':
          description: Pointlist files
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      $ref: '#/components/schemas/PointlistFile'
                  count:
                    type: integer
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/BudgetExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/pointlists/sources:
    get:
      summary: List Pointlist Sources
      description: The imported pointlist series with their coverage, file and point counts.
      operationId: getPointlistSources
      tags:
        - Points
      parameters:
        - $ref: '#/components/parameters/DomainDefaultFidonet'
      responses:
        '200':
          description: Pointlist sources
          content:
            application/json:
              schema:
                type: object
                properties:
                  sources:
                    type: array
                    items:
                      $ref: '#/components/schemas/PointlistSource'
                  count:
                    type: integer
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/BudgetExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/modem/pstn-dead:
    post:
      summary: Mark a PSTN Node Dead
      description: |
        Record that a node's phone number does not answer, so later callers
        skip it. Authenticated; registered only when modem_api is enabled.
        A body whose Content-Length exceeds the configured size limit is
        refused with 413; a chunked body that grows past it fails decoding
        (400).
      operationId: markPSTNDead
      tags:
        - Modem Testing
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [zone, net, node]
              properties:
                zone:
                  type: integer
                net:
                  type: integer
                node:
                  type: integer
                reason:
                  type: string
      responses:
        '200':
          description: Node marked dead
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: Content-Length larger than the configured limit
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Revive a PSTN Node
      description: |
        Clear a node's dead mark so it is dialled again. Authenticated;
        registered only when modem_api is enabled. A body whose
        Content-Length exceeds the configured size limit is refused with
        413; a chunked body that grows past it fails decoding (400).
      operationId: unmarkPSTNDead
      tags:
        - Modem Testing
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [zone, net, node]
              properties:
                zone:
                  type: integer
                net:
                  type: integer
                node:
                  type: integer
      responses:
        '200':
          description: Node revived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: Content-Length larger than the configured limit
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    AddressEnvelope:
      type: object
      description: >
        The preamble of every response about one 3-D or 4-D address: which
        network answered and which networks the address exists in.
      properties:
        address:
          type: string
          example: "2:5001/100"
        domain:
          type: string
          example: fidonet
        available_domains:
          type: array
          items:
            type: string
          description: Always an array, possibly empty
...
    PointSearchFilter:
      type: object
      description: The point search's reading of its parameters (null = not applied)
      properties:
        zone:
          type: integer
          nullable: true
        net:
          type: integer
          nullable: true
        node:
          type: integer
          nullable: true
        point:
          type: integer
          nullable: true
        domain:
          type: string
          nullable: true
        list_source:
          type: string
          nullable: true
        system_name:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        sysop_name:
          type: string
          nullable: true
        date_from:
          type: string
          format: date-time
          nullable: true
        date_to:
          type: string
          format: date-time
          nullable: true
        latest_only:
          type: boolean
          nullable: true
        limit:
          type: integer
        offset:
          type: integer
...
    StatusMessage:
      type: object
      properties:
        status:
          type: string
          example: ok
        message:
          type: string
    Error:
      type: object
      description: Every JSON error body
      properties:
        error:
          type: string
          description: What went wrong
        status:
          type: integer
          description: The HTTP status, repeated
        time:
          type: string
          format: date-time
...
    Point:
      type: object
      properties:
        zone:
          type: integer
        net:
          type: integer
        node:
          type: integer
        point:
          type: integer
        domain:
          type: string
          description: Omitted when empty.
        pointlist_date:
          type: string
          format: date-time
        day_number:
          type: integer
        list_source:
          type: string
        source_priority:
          type: integer
        source_format:
          type: string
        system_name:
          type: string
        location:
          type: string
        sysop_name:
          type: string
        phone:
          type: string
        max_speed:
          type: integer
        is_cm:
          type: boolean
        is_mo:
          type: boolean
        has_inet:
          type: boolean
        flags:
          type: array
          items:
            type: string
        modem_flags:
          type: array
          items:
            type: string
        internet_config:
          type: object
          additionalProperties: true
          description: Omitted when empty.
        conflict_sequence:
          type: integer
        has_conflict:
          type: boolean
        fts_id:
          type: string
        raw_line:
          type: string
          description: Omitted when empty.
    PointlistFile:
      type: object
      properties:
        domain:
          type: string
        list_source:
          type: string
        pointlist_date:
          type: string
          format: date-time
        day_number:
          type: integer
        filename:
          type: string
        source_format:
          type: string
        points_count:
          type: integer
        bosses_count:
          type: integer
        imported_at:
          type: string
          format: date-time
    PointlistSource:
      type: object
      properties:
        domain:
          type: string
        list_source:
          type: string
        first_date:
          type: string
          format: date-time
        last_date:
          type: string
          format: date-time
        file_count:
          type: integer
        total_points:
          type: integer
...
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: |
        A caller-specific API key. The server stores its SHA-256 hash in
        modem_api.callers; see the README for generating a pair.
  responses:
    BadRequest:
      description: Invalid or missing parameter
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: No such record
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: >
        Rate limited. The body is plain text and Retry-After says how many
        seconds to wait.
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        text/plain:
          schema:
            type: string
    BudgetExceeded:
      description: The query outran its time budget; narrow it or retry later
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    Zone:
      name: zone
      in: path
      required: true
      description: FTN zone number (FidoNet uses 1-6; other networks have their own)
      schema:
        type: integer
        minimum: 1
    Net:
      name: net
      in: path
      required: true
      description: Network number (0-65535)
      schema:
        type: integer
        minimum: 0
        maximum: 65535
    Node:
      name: node
      in: path
      required: true
      description: Node number (0-65535)
      schema:
        type: integer
        minimum: 0
        maximum: 65535
    PointNumber:
      name: point
      in: path
      required: true
      description: Point number under the boss
      schema:
        type: integer
        minimum: 0
    DomainResolved:
      name: domain
      in: query
      description: >
        FTN network. Omitted: the only network the address exists in, or
        fidonet when it exists in several.
      schema:
        type: string
        example: fsxnet
    DomainDefaultFidonet:
      name: domain
      in: query
      description: FTN network (defaults to fidonet)
      schema:
        type: string
        default: fidonet
...